refactor, use element in both builders for consistency. nutrition value can be Integer rather than Double.

+++ src/main/java/com/sainsbury/domain/ProductPageInfo.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Setter
 public class ProductPageInfo {
 
-    private Double kcalPer100g;
+    private Integer kcalPer100g;
 
     private String description;
 }
